:truck: Split dropdown_filter_chip.dart into separate classes
